backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
uvicorn
notion-client
redis[hiredis]
plaid-python